requirements.txt
APScheduler==3.10.4
asgiref==3.5.2
cachelib==0.10.2
certifi==2020.12.5
chardet==3.0.4
click==8.1.3
django-widget-tweaks==1.4.12
dominate==2.8.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
flask-session==0.5.0
google==3.0.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.1.0
Pillow==9.2.0
ping3==4.0.4
protobuf==4.21.12
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nmap==0.6.1
soupsieve==2.4
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.7
tzlocal==5.1
urllib3==1.26.2
visitor==0.1.3
Werkzeug==2.2.2
zipp==3.11.0
